change direction of odds ratio calculation
Formerly, if two species were given (sp1 and sp2), the direction of
association was sp1~sp2. In retrospect this didn’t seem sensible, as it
causes problems for later plotting programs (due to need to reorder
columns from spaa, or transpose matrices from make.or.matrix) for the
direction of each association to be plotted correctly.

This change means that asymmetric odds ratio results from this (and all
future) versions of sppairs will not match those from earlier versions.