new functions to ensure consistent transformation between matrix and data.frame
SPAA returns a data.frame, but sometimes it is useful to return these
results as a (square) matrix; consequently it is necessary to be able
to transfer to from ‘long’ to ‘wide’ format (and back again) without
introducing errors. The new functions achieve that, albeit not as
cleanly as they might.