allow functions passed to spaa to return multiple values
Previously, any function passed to spaa had to return a single value,
but this was problematic in general (e.g. where a test returned a
statistic and a p value). This has been corrected so that functions
that return multiple values are treated somewhat more sensibly. Also
misc additions to association functions, updated help etc.